only creator can edit and delete

+++ backend/routes/posts.js
@@ -27,7 +27,7 @@ const storage = multer.diskStorage({
   }
 });
 
-router.post("",checkAuth, multer({ storage: storage }).single("image"), (req, res, next) => {
+router.post("", checkAuth, multer({ storage: storage }).single("image"), (req, res, next) => {
   const url = req.protocol + "://" + req.get("host");
   const post = new Post({
     title: req.body.title,
@@ -55,43 +55,48 @@ router.post("",checkAuth, multer({ storage: storage }).single("image"), (req, re
 
 router.get("", (req, res, next) => {
   //console.log(req.query);
-  const pageSize=+req.query.pageSize;
-  const currentPage=+req.query.currentPage;
-  const postQuery=Post.find();
+  const pageSize = +req.query.pageSize;
+  const currentPage = +req.query.currentPage;
+  const postQuery = Post.find();
   let fetechedPosts;
-  if(pageSize && currentPage){
-    postQuery.skip(pageSize*(currentPage-1)).limit(pageSize);
+  if (pageSize && currentPage) {
+    postQuery.skip(pageSize * (currentPage - 1)).limit(pageSize);
   }
   postQuery.then((documents) => {
     //console.log(documents);
-    fetechedPosts=documents;
+    fetechedPosts = documents;
     return Post.count();
-  }).then((count)=>{
+  }).then((count) => {
     res.status(200).json({
       message: "Posts fetched successfully!",
       posts: fetechedPosts,
-      maxPosts:count,
+      maxPosts: count,
     });
   });
 
 });
 
-router.put("/:id",checkAuth, multer({ storage: storage }).single("image"), (req, res, next) => {
-  let imagePath=req.body.imagePath;
+router.put("/:id", checkAuth, multer({ storage: storage }).single("image"), (req, res, next) => {
+  let imagePath = req.body.imagePath;
   //console.log(imagePath)
   if (req.file) {
     const url = req.protocol + "://" + req.get("host");
-    imagePath= url + "/images/" + req.file.filename
+    imagePath = url + "/images/" + req.file.filename
   }
   const post = new Post({
     _id: req.params.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath:imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  Post.updateOne({ _id: req.params.id }, post).then((result) => {
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then((result) => {
     //console.log(result);
-    res.status(200).json({ message: "Post updated!" });
+    if (result.nModified > 0) {
+      res.status(200).json({ message: "Post updated!" });
+    } else {
+      res.status(401).json({ message: "Not Authorized!" });
+    }
   })
 });
 
@@ -105,11 +110,15 @@ router.get("/:id", (req, res, next) => {
   })
 })
 
-router.delete("/:id",checkAuth, (req, res, next) => {
+router.delete("/:id", checkAuth, (req, res, next) => {
   //console.log(req.params.id);
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
-    //console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator:req.userData.userId }).then((result) => {
+    console.log(result);
+    if (result.n > 0) {
+      res.status(200).json({ message: "Post deleted!" });
+    } else {
+      res.status(401).json({ message: "Not Authorized!" });
+    }
   })
 });
